CLI-784 Split SQAA command into focused modules
Extract context resolution, run orchestration, JSON report building,
and shared types from sqaa.ts into dedicated files to keep the
command entrypoint thin after adding telemetry instrumentation.
